make SRU and attention 2060 compatible

+++ models/hrm/hrm_act_v3.py
@@ -53,7 +53,7 @@ class HierarchicalReasoningModel_ACTV3Config(BaseModel):
     halt_max_steps: int
     halt_exploration_prob: float
 
-    forward_dtype: str = "bfloat16"
+    forward_dtype: str = "float16"
 
 
 class HierarchicalReasoningModel_ACTV3_Inner(nn.Module):
